backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pytest==7.4.3
httpx==0.26.0
python-dotenv==1.0.0
